Rename other g_proximal headers
