+++ cpp/examples/forward_backward/inpainting_joint_map.cc
@@ -7,7 +7,7 @@
 #include <ctime>
 
 #include "sopt/imaging_forward_backward.h"
-#include "sopt/l1_g_proximal.h"
+#include "sopt/l1_non_diff_function.h"
 #include "sopt/joint_map.h"
 #include "sopt/logging.h"
 #include "sopt/maths.h"
