+++ cpp/sopt/l1_non_diff_function.h
