+++ cpp/sopt/tf_non_diff_function.h
