+++ cpp/tests/forward_backward.cc
@@ -5,7 +5,7 @@
 #include <Eigen/Dense>
 
 #include "sopt/imaging_forward_backward.h"
-#include "sopt/l1_g_proximal.h"
+#include "sopt/l1_non_diff_function.h"
 #include "sopt/logging.h"
 #include "sopt/maths.h"
 #include "sopt/proximal.h"
